Rewrite bucket backend to extend resolved str lifetime
Bucket backend will no longer copy and move buckets (as it's supposed to),
which allows resolved strings to be valid for entirety of interner
lifetime ('i).

To do this I replaced fixed_str with correct Bucket implementations
which rely on Pin to guarantee the data won't be moved.

Improved InternedStr as well, which is now safe to use externally
because its validity is enforced by borrow checker.

Updated backend code to reflect these changes.

Also rewrote Clone implementation for bucket backend to avoid nasty
overallocation it was prone to doing previously because a cloned bucket
would have `head` bucket that's as big as all previous allocations
combined - if that allocation didn't fail, the head that's created next
could very likely be too big for allocation.

Added a `with_backend` method to main interner struct which allows
specifying exact backend construction parameters instead of forcing all
backends to rely on heuristics or incorrect/bad assumptions for most
initializations.

Signed-off-by: Tin Švagelj <tin.svagelj@live.com>